refactor(bridge): extract profile pictures
